Split out integration tests mocks into a module
